[SPARK-44392][SQL][TESTS] Add tests for schema pruning in delta-based UPDATEs
