+++ sql/core/src/test/scala/org/apache/spark/sql/connector/DeltaBasedUpdateTableSuite.scala
@@ -17,61 +17,11 @@
 
 package org.apache.spark.sql.connector
 
-import org.apache.spark.sql.{AnalysisException, Row}
-
-class DeltaBasedUpdateTableSuite extends UpdateTableSuiteBase {
+class DeltaBasedUpdateTableSuite extends DeltaBasedUpdateTableSuiteBase {
 
   override protected lazy val extraTableProps: java.util.Map[String, String] = {
     val props = new java.util.HashMap[String, String]()
     props.put("supports-deltas", "true")
     props
-  }
-
-  test("nullable row ID attrs") {
-    createAndInitTable("pk INT, salary INT, dep STRING",
-      """{ "pk": 1, "salary": 300, "dep": 'hr' }
-        |{ "pk": 2, "salary": 150, "dep": 'software' }
-        |{ "pk": 3, "salary": 120, "dep": 'hr' }
-        |""".stripMargin)
-
-    val exception = intercept[AnalysisException] {
-      sql(s"UPDATE $tableNameAsString SET salary = -1 WHERE pk = 1")
-    }
-    assert(exception.message.contains("Row ID attributes cannot be nullable"))
-  }
-
-  test("update with assignments to row ID") {
-    createAndInitTable("pk INT NOT NULL, id INT, dep STRING",
-      """{ "pk": 1, "id": 1, "dep": "hr" }
-        |{ "pk": 2, "id": 2, "dep": "software" }
-        |{ "pk": 3, "id": 3, "dep": "hr" }
-        |""".stripMargin)
-
-    sql(s"UPDATE $tableNameAsString SET pk = 10 WHERE id = 1")
-
-    checkAnswer(
-      sql(s"SELECT * FROM $tableNameAsString"),
-      Row(10, 1, "hr") :: Row(2, 2, "software") :: Row(3, 3, "hr") :: Nil)
-  }
-
-  test("update with nondeterministic conditions") {
-    createAndInitTable("pk INT NOT NULL, id INT, dep STRING",
-      """{ "pk": 1, "id": 1, "dep": "hr" }
-        |{ "pk": 2, "id": 2, "dep": "software" }
-        |{ "pk": 3, "id": 3, "dep": "hr" }
-        |""".stripMargin)
-
-    checkError(
-      exception = intercept[AnalysisException] {
-        sql(s"UPDATE $tableNameAsString SET dep = 'invalid' WHERE id <= 1 AND rand() > 0.5")
-      },
-      errorClass = "INVALID_NON_DETERMINISTIC_EXPRESSIONS",
-      parameters = Map(
-        "sqlExprs" -> "\"((id <= 1) AND (rand() > 0.5))\""),
-      context = ExpectedContext(
-        fragment = "UPDATE cat.ns1.test_table SET dep = 'invalid' WHERE id <= 1 AND rand() > 0.5",
-        start = 0,
-        stop = 75)
-    )
   }
 }
